Add dummy data to stories
