+++ src/components/IDO/IDO.tsx
@@ -2,53 +2,24 @@ import * as React from 'react'
 import { Grid, Typography, Box, CardMedia } from '@material-ui/core'
 import useStyles from './style'
 import ExchangeAmountInput from '@components/Inputs/ExchangeAmountInput/ExchangeAmountInput'
-import { PublicKey } from '@solana/web3.js'
-import { BN } from '@project-serum/anchor'
-import { SwapToken } from '@components/Swap/Swap'
+
 import { OutlinedButton } from '@components/OutlinedButton/OutlinedButton'
 import watchIcon from '@static/svg/watch.svg'
 import solanaIcon from '@static/svg/solanaToken.svg'
 import dollarIcon from '@static/svg/dollarToken.svg'
 import invariantIcon from '@static/svg/invariantToken.svg'
 import invariantLogo from '@static/svg/invariantLogo.svg'
-const IDO: React.FC = () => {
+import { SwapToken } from '@components/Swap/Swap'
+
+export interface IIDO {
+  tokens: SwapToken[]
+}
+
+const IDO: React.FC<IIDO> = ({ tokens }) => {
   const classes = useStyles()
   const getButtonMessage = () => {
     return 'Connect a wallet'
   }
-
-  const tokens: SwapToken[] = [
-    {
-      balance: new BN(100).mul(new BN(34786)),
-      decimal: 6,
-      symbol: 'SOL',
-      assetAddress: new PublicKey('So11111111111111111111111111111111111111112'),
-      name: 'Wrapped Solana',
-      logoURI:
-        'https://raw.githubusercontent.com/solana-labs/token-list/main/assets/mainnet/So11111111111111111111111111111111111111112/logo.png',
-      address: new PublicKey('So11111111111111111111111111111111111111112')
-    },
-    {
-      balance: new BN(100).mul(new BN(126)),
-      decimal: 6,
-      symbol: 'BTC',
-      assetAddress: new PublicKey('9n4nbM75f5Ui33ZbPYXn59EwSgE8CGsHtAeTH5YFeJ9E'),
-      name: 'BTC',
-      logoURI:
-        'https://raw.githubusercontent.com/solana-labs/token-list/main/assets/mainnet/9n4nbM75f5Ui33ZbPYXn59EwSgE8CGsHtAeTH5YFeJ9E/logo.png',
-      address: new PublicKey('So11111111111111111111111111111111111111112')
-    },
-    {
-      balance: new BN(10).mul(new BN(5342)),
-      decimal: 6,
-      symbol: 'USDC',
-      assetAddress: new PublicKey('EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v'),
-      name: 'USD coin',
-      logoURI:
-        'https://raw.githubusercontent.com/solana-labs/token-list/main/assets/mainnet/EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v/logo.png',
-      address: new PublicKey('EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v')
-    }
-  ]
 
   const [tokenFromIndex, setTokenFromIndex] = React.useState<number | null>(
     tokens.length ? 0 : null
